Convert edit comment into a dropdown item
